2018_12_11

+++ libft/ft_memcpy.c
@@ -3,10 +3,10 @@
 /*                                                        :::      ::::::::   */
 /*   ft_memcpy.c                                        :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
-/*   By: creek <marvin@42.fr>                       +#+  +:+       +#+        */
+/*   By: creek <creek@student.42.fr>                +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2018/11/25 23:01:30 by creek             #+#    #+#             */
-/*   Updated: 2018/11/26 15:48:38 by creek            ###   ########.fr       */
+/*   Updated: 2018/12/06 17:38:32 by creek            ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -14,13 +14,13 @@
 
 void	*ft_memcpy(void *dst, const void *src, size_t n)
 {
-	int i;
-	char *new_dst;
-	char *new_src;
+	size_t	i;
+	char	*new_dst;
+	char	*new_src;
 
 	i = 0;
-	new_dst = (char *) dst;
-	new_src = (char *) src;
+	new_dst = (char *)dst;
+	new_src = (char *)src;
 	while (i < n)
 	{
 		new_dst[i] = new_src[i];
