+++ libft/ft_strncat.c
@@ -3,10 +3,10 @@
 /*                                                        :::      ::::::::   */
 /*   ft_strncat.c                                       :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
-/*   By: creek <marvin@42.fr>                       +#+  +:+       +#+        */
+/*   By: creek <creek@student.42.fr>                +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2018/11/27 19:02:30 by creek             #+#    #+#             */
-/*   Updated: 2018/11/28 19:06:55 by creek            ###   ########.fr       */
+/*   Updated: 2018/12/06 17:41:10 by creek            ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -14,9 +14,9 @@
 
 char	*ft_strncat(char *s1, const char *s2, size_t n)
 {
-	int i;
-	int j;
-	
+	size_t	i;
+	int		j;
+
 	i = 0;
 	j = 0;
 	while (s1[j])
